move some functions to utils to reuse from other project
