+++ examples/nlp/text_normalization_as_tagging/dataset_preparation/extract_giza_alignments.py
@@ -20,8 +20,10 @@ This script can be used after GIZA++ alignment to extract final alignments for e
 import re
 from argparse import ArgumentParser
 from typing import List, Tuple
+from nemo.collections.nlp.data.spellchecking_asr_customization.utils import (
+    fill_alignment_matrix, check_monotonicity, get_targets, get_targets_from_back
+)
 
-import numpy as np
 
 parser = ArgumentParser(description='Extract final alignments from GIZA++ alignments')
 parser.add_argument('--mode', type=str, required=True, help='tn or itn')
@@ -34,96 +36,7 @@ parser.add_argument('--lang', type=str, required=True, help="Language")
 args = parser.parse_args()
 
 
-def fill_alignment_matrix(
-    fline2: str, fline3: str, gline2: str, gline3: str
-) -> Tuple[np.ndarray, List[str], List[str]]:
-    """Parse Giza++ direct and reverse alignment results and represent them as an alignment matrix
-
-    Args:
-        fline2: e.g. "_2 0 1 4_"
-        fline3: e.g. "NULL ({ }) twenty ({ 1 }) fourteen ({ 2 3 4 })"
-        gline2: e.g. "twenty fourteen"
-        gline3: e.g. "NULL ({ }) _2 ({ 1 }) 0 ({ }) 1 ({ }) 4_ ({ 2 })"
-
-    Returns:
-        matrix: a numpy array of shape (src_len, dst_len) filled with [0, 1, 2, 3], where 3 means a reliable alignment
-         the corresponding words were aligned to one another in direct and reverse alignment runs, 1 and 2 mean that the
-         words were aligned only in one direction, 0 - no alignment.
-        srctokens: e.g. ["twenty", "fourteen"]
-        dsttokens: e.g. ["_2", "0", "1", "4_"]
-
-    For example, the alignment matrix for the above example may look like:
-    [[3, 0, 0, 0]
-     [0, 2, 2, 3]]
-    """
-    if fline2 is None or gline2 is None or fline3 is None or gline3 is None:
-        raise ValueError(f"empty params")
-    srctokens = gline2.split()
-    dsttokens = fline2.split()
-    pattern = r"([^ ]+) \(\{ ([^\(\{\}\)]*) \}\)"
-    src2dst = re.findall(pattern, fline3.replace("({ })", "({  })"))
-    dst2src = re.findall(pattern, gline3.replace("({ })", "({  })"))
-    if len(src2dst) != len(srctokens) + 1:
-        raise ValueError(
-            "length mismatch: len(src2dst)="
-            + str(len(src2dst))
-            + "; len(srctokens)"
-            + str(len(srctokens))
-            + "\n"
-            + gline2
-            + "\n"
-            + fline3
-        )
-    if len(dst2src) != len(dsttokens) + 1:
-        raise ValueError(
-            "length mismatch: len(dst2src)="
-            + str(len(dst2src))
-            + "; len(dsttokens)"
-            + str(len(dsttokens))
-            + "\n"
-            + fline2
-            + "\n"
-            + gline3
-        )
-    matrix = np.zeros((len(srctokens), len(dsttokens)))
-    for i in range(1, len(src2dst)):
-        token, to_str = src2dst[i]
-        if to_str == "":
-            continue
-        to = list(map(int, to_str.split()))
-        for t in to:
-            matrix[i - 1][t - 1] = 2
-
-    for i in range(1, len(dst2src)):
-        token, to_str = dst2src[i]
-        if to_str == "":
-            continue
-        to = list(map(int, to_str.split()))
-        for t in to:
-            matrix[t - 1][i - 1] += 1
-
-    return matrix, srctokens, dsttokens
-
-
-def check_monotonicity(matrix: np.ndarray) -> bool:
-    """Check if alignment is monotonous - i.e. the relative order is preserved (no swaps).
-
-    Args:
-        matrix: a numpy array of shape (src_len, dst_len) filled with [0, 1, 2, 3], where 3 means a reliable alignment
-         the corresponding words were aligned to one another in direct and reverse alignment runs, 1 and 2 mean that the
-         words were aligned only in one direction, 0 - no alignment.
-    """
-    is_sorted = lambda k: np.all(k[:-1] <= k[1:])
-
-    a = np.argwhere(matrix == 3)
-    b = np.argwhere(matrix == 2)
-    c = np.vstack((a, b))
-    d = c[c[:, 1].argsort()]  # sort by second column (less important)
-    d = d[d[:, 0].argsort(kind="mergesort")]
-    return is_sorted(d[:, 1])
-
-
-def get_targets(matrix: np.ndarray, dsttokens: List[str]) -> List[str]:
+def get_targets(matrix: np.ndarray, dsttokens: List[str], delimiter: str) -> List[str]:
     """Join some of the destination tokens, so that their number becomes the same as the number of input words.
     Unaligned tokens tend to join to the left aligned token.
 
@@ -171,16 +84,13 @@ def get_targets(matrix: np.ndarray, dsttokens: List[str]) -> List[str]:
                         break
 
         if len(dstlist) > 0:
-            if args.mode == "tn":
-                targets.append("_".join(dstlist))
-            else:
-                targets.append("".join(dstlist))
+            targets.append(delimiter.join(dstlist))
         else:
             targets.append("<DELETE>")
     return targets
 
 
-def get_targets_from_back(matrix: np.ndarray, dsttokens: List[str]) -> List[str]:
+def get_targets_from_back(matrix: np.ndarray, dsttokens: List[str], delimiter: str) -> List[str]:
     """Join some of the destination tokens, so that their number becomes the same as the number of input words.
     Unaligned tokens tend to join to the right aligned token.
 
@@ -225,10 +135,7 @@ def get_targets_from_back(matrix: np.ndarray, dsttokens: List[str]) -> List[str]
                     else:
                         break
         if len(dstlist) > 0:
-            if args.mode == "tn":
-                targets.append("_".join(list(reversed(dstlist))))
-            else:
-                targets.append("".join(list(reversed(dstlist))))
+            targets.append(delimiter.join(list(reversed(dstlist))))
         else:
             targets.append("<DELETE>")
     return list(reversed(targets))
@@ -237,8 +144,10 @@ def get_targets_from_back(matrix: np.ndarray, dsttokens: List[str]) -> List[str]
 def main() -> None:
     g = open(args.giza_dir + "/GIZA++." + args.giza_suffix, "r", encoding="utf-8")
     f = open(args.giza_dir + "/GIZA++reverse." + args.giza_suffix, "r", encoding="utf-8")
+    target_inner_delimiter = ""
     if args.mode == "tn":
         g, f = f, g
+        target_inner_delimiter = "_"
     out = open(args.giza_dir + "/" + args.out_filename, "w", encoding="utf-8")
     cache = {}
     good_count, not_mono_count, not_covered_count, exception_count = 0, 0, 0, 0
@@ -277,8 +186,8 @@ def main() -> None:
         else:
             matrix[matrix <= 2] = 0  # leave only 1-to-1 alignment points
             if check_monotonicity(matrix):
-                targets = get_targets(matrix, dsttokens)
-                targets_from_back = get_targets_from_back(matrix, dsttokens)
+                targets = get_targets(matrix, dsttokens, delimiter=target_inner_delimiter)
+                targets_from_back = get_targets_from_back(matrix, dsttokens, delimiter=target_inner_delimiter)
                 if len(targets) != len(srctokens):
                     raise ValueError(
                         "targets length doesn't match srctokens length: len(targets)="
