+++ nemo/collections/nlp/data/text_normalization_as_tagging/utils.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 
+import numpy as np
 import re
 from itertools import groupby
 from typing import Dict, List, Tuple
@@ -305,3 +306,197 @@ def get_src_and_dst_for_alignment(
         )
 
     return written_str, spoken, " ".join(same_begin), " ".join(same_end)
+
+
+def fill_alignment_matrix(
+    fline2: str, fline3: str, gline2: str, gline3: str
+) -> Tuple[np.ndarray, List[str], List[str]]:
+    """Parse Giza++ direct and reverse alignment results and represent them as an alignment matrix
+
+    Args:
+        fline2: e.g. "_2 0 1 4_"
+        fline3: e.g. "NULL ({ }) twenty ({ 1 }) fourteen ({ 2 3 4 })"
+        gline2: e.g. "twenty fourteen"
+        gline3: e.g. "NULL ({ }) _2 ({ 1 }) 0 ({ }) 1 ({ }) 4_ ({ 2 })"
+
+    Returns:
+        matrix: a numpy array of shape (src_len, dst_len) filled with [0, 1, 2, 3], where 3 means a reliable alignment
+         the corresponding words were aligned to one another in direct and reverse alignment runs, 1 and 2 mean that the
+         words were aligned only in one direction, 0 - no alignment.
+        srctokens: e.g. ["twenty", "fourteen"]
+        dsttokens: e.g. ["_2", "0", "1", "4_"]
+
+    For example, the alignment matrix for the above example may look like:
+    [[3, 0, 0, 0]
+     [0, 2, 2, 3]]
+    """
+    if fline2 is None or gline2 is None or fline3 is None or gline3 is None:
+        raise ValueError(f"empty params")
+    srctokens = gline2.split()
+    dsttokens = fline2.split()
+    pattern = r"([^ ]+) \(\{ ([^\(\{\}\)]*) \}\)"
+    src2dst = re.findall(pattern, fline3.replace("({ })", "({  })"))
+    dst2src = re.findall(pattern, gline3.replace("({ })", "({  })"))
+    if len(src2dst) != len(srctokens) + 1:
+        raise ValueError(
+            "length mismatch: len(src2dst)="
+            + str(len(src2dst))
+            + "; len(srctokens)"
+            + str(len(srctokens))
+            + "\n"
+            + gline2
+            + "\n"
+            + fline3
+        )
+    if len(dst2src) != len(dsttokens) + 1:
+        raise ValueError(
+            "length mismatch: len(dst2src)="
+            + str(len(dst2src))
+            + "; len(dsttokens)"
+            + str(len(dsttokens))
+            + "\n"
+            + fline2
+            + "\n"
+            + gline3
+        )
+    matrix = np.zeros((len(srctokens), len(dsttokens)))
+    for i in range(1, len(src2dst)):
+        token, to_str = src2dst[i]
+        if to_str == "":
+            continue
+        to = list(map(int, to_str.split()))
+        for t in to:
+            matrix[i - 1][t - 1] = 2
+
+    for i in range(1, len(dst2src)):
+        token, to_str = dst2src[i]
+        if to_str == "":
+            continue
+        to = list(map(int, to_str.split()))
+        for t in to:
+            matrix[t - 1][i - 1] += 1
+
+    return matrix, srctokens, dsttokens
+
+
+def check_monotonicity(matrix: np.ndarray) -> bool:
+    """Check if alignment is monotonous - i.e. the relative order is preserved (no swaps).
+
+    Args:
+        matrix: a numpy array of shape (src_len, dst_len) filled with [0, 1, 2, 3], where 3 means a reliable alignment
+         the corresponding words were aligned to one another in direct and reverse alignment runs, 1 and 2 mean that the
+         words were aligned only in one direction, 0 - no alignment.
+    """
+    is_sorted = lambda k: np.all(k[:-1] <= k[1:])
+
+    a = np.argwhere(matrix == 3)
+    b = np.argwhere(matrix == 2)
+    c = np.vstack((a, b))
+    d = c[c[:, 1].argsort()]  # sort by second column (less important)
+    d = d[d[:, 0].argsort(kind="mergesort")]
+    return is_sorted(d[:, 1])
+
+
+def get_targets(matrix: np.ndarray, dsttokens: List[str], delimiter: str) -> List[str]:
+    """Join some of the destination tokens, so that their number becomes the same as the number of input words.
+    Unaligned tokens tend to join to the left aligned token.
+
+    Args:
+        matrix: a numpy array of shape (src_len, dst_len) filled with [0, 1, 2, 3], where 3 means a reliable alignment
+         the corresponding words were aligned to one another in direct and reverse alignment runs, 1 and 2 mean that the
+         words were aligned only in one direction, 0 - no alignment.
+        dsttokens: e.g. ["_2", "0", "1", "4_"]
+    Returns:
+        targets: list of string tokens, with one-to-one correspondence to matrix.shape[0]
+
+    Example:
+        If we get
+            matrix=[[3, 0, 0, 0]
+                    [0, 2, 2, 3]]
+            dsttokens=["_2", "0", "1", "4_"]
+        it gives
+            targets = ["_201", "4_"]
+        Actually, this is a mistake instead of ["_20", "14_"]. That will be further corrected by regular expressions.
+    """
+    targets = []
+    last_covered_dst_id = -1
+    for i in range(len(matrix)):
+        dstlist = []
+        for j in range(last_covered_dst_id + 1, len(dsttokens)):
+            # matrix[i][j] == 3: safe alignment point
+            if matrix[i][j] == 3 or (
+                j == last_covered_dst_id + 1
+                and np.all(matrix[i, :] == 0)  # if the whole line does not have safe points
+                and np.all(matrix[:, j] == 0)  # and the whole column does not have safe points, match them
+            ):
+                if len(targets) == 0:  # if this is first safe point, attach left unaligned columns to it, if any
+                    for k in range(0, j):
+                        if np.all(matrix[:, k] == 0):  # if column k does not have safe points
+                            dstlist.append(dsttokens[k])
+                        else:
+                            break
+                dstlist.append(dsttokens[j])
+                last_covered_dst_id = j
+                for k in range(j + 1, len(dsttokens)):
+                    if np.all(matrix[:, k] == 0):  # if column k does not have safe points
+                        dstlist.append(dsttokens[k])
+                        last_covered_dst_id = k
+                    else:
+                        break
+
+        if len(dstlist) > 0:
+            targets.append(delimiter.join(dstlist))
+        else:
+            targets.append("<DELETE>")
+    return targets
+
+
+def get_targets_from_back(matrix: np.ndarray, dsttokens: List[str], delimiter: str) -> List[str]:
+    """Join some of the destination tokens, so that their number becomes the same as the number of input words.
+    Unaligned tokens tend to join to the right aligned token.
+
+    Args:
+        matrix: a numpy array of shape (src_len, dst_len) filled with [0, 1, 2, 3], where 3 means a reliable alignment
+         the corresponding words were aligned to one another in direct and reverse alignment runs, 1 and 2 mean that the
+         words were aligned only in one direction, 0 - no alignment.
+        dsttokens: e.g. ["_2", "0", "1", "4_"]
+    Returns:
+        targets: list of string tokens, with one-to-one correspondence to matrix.shape[0]
+
+    Example:
+        If we get
+            matrix=[[3, 0, 0, 0]
+                    [0, 2, 2, 3]]
+            dsttokens=["_2", "0", "1", "4_"]
+        it gives
+            targets = ["_2", "014_"]
+        Actually, this is a mistake instead of ["_20", "14_"]. That will be further corrected by regular expressions.
+    """
+
+    targets = []
+    last_covered_dst_id = len(dsttokens)
+    for i in range(len(matrix) - 1, -1, -1):
+        dstlist = []
+        for j in range(last_covered_dst_id - 1, -1, -1):
+            if matrix[i][j] == 3 or (
+                j == last_covered_dst_id - 1 and np.all(matrix[i, :] == 0) and np.all(matrix[:, j] == 0)
+            ):
+                if len(targets) == 0:
+                    for k in range(len(dsttokens) - 1, j, -1):
+                        if np.all(matrix[:, k] == 0):
+                            dstlist.append(dsttokens[k])
+                        else:
+                            break
+                dstlist.append(dsttokens[j])
+                last_covered_dst_id = j
+                for k in range(j - 1, -1, -1):
+                    if np.all(matrix[:, k] == 0):
+                        dstlist.append(dsttokens[k])
+                        last_covered_dst_id = k
+                    else:
+                        break
+        if len(dstlist) > 0:
+            targets.append(delimiter.join(list(reversed(dstlist))))
+        else:
+            targets.append("<DELETE>")
+    return list(reversed(targets))
